Kaelvikkenna Badhil Quiz Tool - March 2021 build
What can this do...?
This application is capable of creating MCQ quizzes that can be answered by (unique) respondents.
The responses are automatically graded & the scores are displayed to the respondent, immediately upon the completion of the test.
The scores of respondents are stored in a database, and  can later be retrieved using the quiz locator.

Kindly go through the "Greetings! Illustrative Guide.pdf", before you proceed...
It might throw some light on how to utilize this tool for its intended purpose...

Built by HARISH KS | March 2021 | IT | 2019-23 PSG Tech
Mmm... It's Ok... Thank me later! ;)

Happy Quizzing!
~ Harish KS, Creator of RCB, SAT & KBQT